remove src/project.ts
