+++ src/core/index.ts
@@ -1,18 +1,20 @@
-export { Channel } from '@/core/channel';
+export { Channel, ChannelType } from '@/core/channel';
 export { EffectConstrains } from '@/core/filters/bounds';
 export * from '@/core/filters/effects';
-export { Track } from '@/core/track';
-export { Playlist } from '@/core/playlist';
+export { Track, TrackType } from '@/core/track';
+export { Playlist, PlaylistType } from '@/core/playlist';
 export { Point } from '@/core/automation/point';
 export { Schedulable } from '@/core/scheduled/schedulable';
 export { ScheduledSample } from '@/core/scheduled/sample';
 export { ScheduledAutomation } from '@/core/scheduled/automation';
-export { Sample } from '@/core/sample';
+export { Sample, SampleType } from '@/core/sample';
 export * from '@/core/automation';
 export { Score } from '@/core/score';
 export { Note } from '@/core/scheduled/note';
 export { Instrument } from '@/core/instrument/instrument';
-export { Pattern } from '@/core/pattern';
+export { Pattern, PatternType } from '@/core/pattern';
 export { ScheduledPattern } from '@/core/scheduled/pattern';
-export { AnyEffect } from '@/core/filters/effect';
+export { AnyEffect, Effect } from '@/core/filters/effect';
 export { Sequence } from '@/core/scheduled/sequence';
+export { SoundfontType, Soundfont } from '@/core/instrument/soundfont';
+export { SynthType, Synth } from '@/core/instrument/synth';
