+++ src/dawg/extensions/core/midi/index.ts
@@ -60,7 +60,7 @@ export const extension: Extension = {
       if (selectedScore.value) {
         const note = new Note(selectedScore.value.instrument, {
           row: keyLookup[e.note.name + e.note.octave].id,
-          duration: noteDuration / 1000 / 60 * project.project.bpm,
+          duration: noteDuration / 1000 / 60 * project.bpm.value,
           time: noteStartTime,
           velocity: noteOn.rawVelocity,
         });
