+++ src/dawg/extensions/core/playlist/index.ts
@@ -59,7 +59,7 @@ export const playlist = manager.activate({
         masterStart,
         masterEnd,
         ghosts,
-        project: project.project,
+        project,
         playlistRowHeight,
         playlistBeatWidth,
       }),
@@ -78,12 +78,12 @@ export const playlist = manager.activate({
           }
 
           const sample = prototype.sample;
-          if (project.project.samples.indexOf(prototype.sample) >= 0) {
+          if (project.samples.indexOf(prototype.sample) >= 0) {
             return;
           }
 
           logger.debug('Adding a sample!');
-          project.project.addSample(sample);
+          project.addSample(sample);
         },
       },
     });
